requirements.txt
matplotlib-base
cartopy
numpy
python==3.9
scipy
numba
obspy==1.2.2
pyqt==5.15.7
pyqtgraph==0.13.1
pandas
cx_oracle
scikit-learn
ipython
joblib
pygments
pexpect
prompt_toolkit
jedi
parso
sphinx
sphinx_rtd_theme
pandoc
jupyterlab